trying with 2 auth pubkeys

+++ programs/clone/src/instructions/add_collateral_to_borrow.rs
@@ -26,7 +26,7 @@ pub struct AddCollateralToBorrow<'info> {
     #[account(
         mut,
         has_one = clone,
-        constraint = token_data.load()?.collaterals[borrow_positions.load()?.borrow_positions[borrow_index as usize].collateral_index as usize].status == Status::Active as u8 @ CloneError::PoolStatusPreventsAction
+        constraint = token_data.load()?.collaterals[borrow_positions.load()?.borrow_positions[borrow_index as usize].collateral_index as usize].status == Status::Active as u64 @ CloneError::PoolStatusPreventsAction
     )]
     pub token_data: AccountLoader<'info, TokenData>,
     #[account(
