+++ programs/clone/src/instructions/swap.rs
@@ -29,7 +29,7 @@ pub struct Swap<'info> {
         mut,
         has_one = clone,
         constraint = (pool_index as u64) < token_data.load()?.num_pools @ CloneError::InvalidInputPositionIndex,
-        constraint = token_data.load()?.pools[pool_index as usize].status == Status::Active as u8 @ CloneError::PoolStatusPreventsAction,
+        constraint = token_data.load()?.pools[pool_index as usize].status == Status::Active as u64 @ CloneError::PoolStatusPreventsAction,
     )]
     pub token_data: AccountLoader<'info, TokenData>,
     #[account(
